requirements.txt
scikit-learn
pandas
psycopg2-binary
sqlalchemy
streamlit
matplotlib
reportlab
